src/backend/app/modules/vision/scanner.py
import base64
import cv2
import numpy as np


class PlotScanner:
    def process_image(self, image_bytes: bytes):
        nparr = np.frombuffer(image_bytes, np.uint8)
        img = cv2.imdecode(nparr, cv2.IMREAD_COLOR)
        img = cv2.resize(img, (800, 600))

        gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
        blur = cv2.GaussianBlur(gray, (5, 5), 0)
        edges = cv2.Canny(blur, 50, 150)

        _, buffer = cv2.imencode(".png", edges)
        edges_base64 = base64.b64encode(buffer).decode("utf-8")

        contours, _ = cv2.findContours(edges, cv2.RETR_LIST, cv2.CHAIN_APPROX_SIMPLE)
        contours = sorted(contours, key=cv2.contourArea, reverse=True)[:5]

        detected_area_px = 0.0
        for contour in contours:
            perimeter = cv2.arcLength(contour, True)
            approx = cv2.approxPolyDP(contour, 0.02 * perimeter, True)
            if len(approx) == 4:
                detected_area_px = cv2.contourArea(contour)
                break

        if detected_area_px < 1000:
            result_stats = {
                "width_ft": 30.0,
                "length_ft": 40.0,
                "area_sqft": 1200.0,
                "confidence": 0.65,
                "status": "approximated",
            }
        else:
            scale_factor = 1200.0 / 50000.0
            real_area = max(600.0, detected_area_px * scale_factor)
            width = (real_area * 0.75) ** 0.5
            length = real_area / width
            result_stats = {
                "width_ft": round(width, 2),
                "length_ft": round(length, 2),
                "area_sqft": round(real_area, 2),
                "confidence": 0.89,
                "status": "detected",
            }

        result_stats["debug_xray"] = edges_base64
        return result_stats
