requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
opencv-python-headless>=4.9.0.80
numpy>=1.26.0
pandas>=2.2.0
streamlit>=1.31.0
requests>=2.31.0
plotly>=5.18.0
pillow>=10.2.0
